Wait for the admin to setup Magento
When multiple nodes start running the configure at the same time, we
will see a collision in the setup scripts running at the same time. This
is a temporary fix to have the "worker" nodes wait a little to let the
admin node finish up initial DB setup. This is purely a workaround until
we have a solution for #90